Hotfixes: - Fix for reading correct ref_addr in dwarf v2. - Workaround for only reading partial decl_files list.

+++ include/orc/dwarf_structs.hpp
@@ -81,7 +81,7 @@ struct attribute_value {
         return _string.hash();
     }
 
-    void reference(std::uint32_t offset) {
+    void reference(std::uint64_t offset) {
         _type |= type::reference;
         _uint = offset;
     }
